requirements.txt
delegator.py==0.1.1
mysqlclient==2.1.1
pexpect==4.8.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
